+++ src/forum/config/TopicsController.java
@@ -14,6 +14,7 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.servlet.ModelAndView;
 
 import forum.*;
+import forum.enums.Permissions;
 
 @Controller
 public class TopicsController {
@@ -21,11 +22,20 @@ public class TopicsController {
 	private TopicService topicService;
 	@Autowired
 	private SectionService sectionService;
+	@Autowired
+	private UserService userService;
 	
 	@SuppressWarnings("unchecked")
 	@RequestMapping(value = "/topics")
-	public ModelAndView topics(@RequestParam("section_id") int id) {
+	public ModelAndView topics(@RequestParam("section_id") int id, @RequestParam(value="error_type", required=false) String error) {
 		ModelAndView model = new ModelAndView("topics");
+		if (error != null) {
+			if (error.equals("bad_permissions")) {
+				model.addObject("delete_topic_error", "Недостаточно прав");
+			} else if (error.equals("existing_topic")) {
+				model.addObject("add_topic_error", "Тема с таким именем уже существует");
+			}
+		}
 		Section section = sectionService.findById(id);
 		List<Topic> topics = new ArrayList<Topic>();
 		section.getTopics().forEach(topic -> topics.add((Topic) topic));
@@ -38,12 +48,21 @@ public class TopicsController {
 	@RequestMapping(value= "/topics/add_topic", method = RequestMethod.POST)
 	public String addTopic(@RequestParam("section_id") int id, @RequestParam("title") String title) {
 		Topic t = new Topic(sectionService.findById(id), title);
-		this.topicService.addTopic(t);
+		try {
+			this.topicService.addTopic(t);
+		} catch (RuntimeException ex) {
+			return "redirect:/topics?section_id=" + id + "&error_type=existing_topic";
+		}
 		return "redirect:/topics?section_id=" + String.valueOf(id);
 	}
 	
 	@RequestMapping(value="/topics/delete_topic")
-	public String deleteTopic(@RequestParam("section_id") int section_id, @RequestParam("topic_id") int topic_id) {
+	public String deleteTopic(@RequestParam("section_id") int section_id, @RequestParam("topic_id") int topic_id,
+			@RequestParam("login") String login) {
+		User user = userService.findByLogin(login);
+		if (user.getPermissions() == Permissions.USER) {
+			return "redirect:/topics?section_id=" + section_id + "&error_type=bad_permissions";
+		}
 		topicService.deleteById(topic_id);
 		return "redirect:/topics?section_id=" + section_id;       
 	}
